backend/requirements.txt
flask
flask-cors
gTTS
python-dotenv
elevenlabs
